Migrate rest package, clean up debug logs
