+++ db/db_test.go
@@ -1,9 +1,11 @@
 package db
 
 import (
+	"context"
 	"log"
 	"testing"
 
+	"github.com/jackc/pgx/v4"
 	"github.com/jackc/pgx/v4/pgxpool"
 	"github.com/likecoin/likecoin-chain-tx-indexer/logger"
 	"go.uber.org/zap/zapcore"
@@ -35,4 +37,21 @@ func TestMain(m *testing.M) {
 
 	InitDB(conn)
 	m.Run()
+}
+
+func debugSQL(tx pgx.Tx, ctx context.Context, sql string, args ...interface{}) (err error) {
+	// add this line to debug SQL (only in test)
+	// debugSQL(tx, ctx, sql, eventStrings, queryString, keywordString, pagination.getOffset(), pagination.Limit)
+	rows, err := tx.Query(ctx, "EXPLAIN "+sql, args...)
+	if err != nil {
+		return err
+	}
+	defer rows.Close()
+
+	for rows.Next() && err == nil {
+		var line string
+		err = rows.Scan(&line)
+		log.Println(line)
+	}
+	return err
 }
